MAJOR ERROR FIX (view description)
An strange unknown error occurred where the form designer said line 416 had an error. After commenting out this line the error moved to 415 and this happened all the way until we couldn't comment out the next line so we deleted everything and downloaded an earlier version of the project and updated it to be the same and it fixed it.